+++ packages/youtube_player_iframe/lib/src/controller/webview_platform_io.dart
@@ -6,7 +6,9 @@ import 'package:webview_flutter/webview_flutter.dart';
 import 'package:webview_flutter_android/webview_flutter_android.dart';
 import 'package:webview_flutter_wkwebview/webview_flutter_wkwebview.dart';
 
-PlatformWebViewControllerCreationParams buildWebViewParams() {
+PlatformWebViewControllerCreationParams buildWebViewParams({
+  bool credentialless = false,
+}) {
   if (WebViewPlatform.instance is WebKitWebViewPlatform) {
     return WebKitWebViewControllerCreationParams(
       allowsInlineMediaPlayback: true,