+++ packages/youtube_player_iframe_web/lib/youtube_player_iframe_web.dart
@@ -1,3 +1,6 @@
 // Copyright 2022 Sarbagya Dhaubanjar. All rights reserved.
 // Use of this source code is governed by a BSD-3-Clause license that can be
 // found in the LICENSE file.
+
+export 'src/web_youtube_player_iframe_controller.dart'
+    show WebYoutubePlayerIframeControllerCreationParams;